Resolve numerical error in output -do_checks
When using ramps, sometimes numerical erros appear causing "out-of-bounds"-errors in do_checks.
We introduce an epsilon=1e-10 in which numerical precision is allowed to vary without detecting an "out-of-bound"-condition. If, due to numerical reasons, the value is out of limits, the instruction value is set to equal the limit value.